+++ InventoryTools/Compendium/Types/MountCompendiumType.cs
@@ -13,6 +13,7 @@ using InventoryTools.Compendium.Interfaces;
 using InventoryTools.Compendium.Models;
 using InventoryTools.Compendium.Sections.Options;
 using InventoryTools.Compendium.Services;
+using InventoryTools.Ui;
 using Lumina.Excel;
 using Lumina.Excel.Sheets;
 
@@ -206,6 +207,7 @@ public class MountCompendiumType : CompendiumType<Mount>
         {
             viewBuilder.AddSingleRowRefSection(new SingleRowRefSectionOptions()
             {
+                SectionKey = "related_item",
                 SectionName = "Related Item",
                 RelatedRef = relatedItem.Value.AsUntypedRowRef()
             });
@@ -215,8 +217,10 @@ public class MountCompendiumType : CompendiumType<Mount>
 
             viewBuilder.AddItemSourcesSection(new ItemSourcesSectionOptions()
                 {
+                    SectionKey = "sources",
                     SectionName = "Sources",
-                    Sources = sources ?? []
+                    Sources = sources ?? [],
+                    SourceType = SourceType.Source
                 });
         }
 
@@ -224,6 +228,7 @@ public class MountCompendiumType : CompendiumType<Mount>
         {
             viewBuilder.AddSingleRowRefSection(new SingleRowRefSectionOptions()
             {
+                SectionKey = "bgm",
                 SectionName = "BGM",
                 RelatedRef = row.RideBGM.Value.AsUntypedRowRef(),
             });