+++ InventoryTools/Extensions/SettingSubCategoryExtension.cs
@@ -63,6 +63,10 @@ namespace InventoryTools.Extensions
                     return "Acquisition Tracker";
                 case SettingSubCategory.IngredientPatch:
                     return "Ingredient Patch";
+                case SettingSubCategory.GlamourReadySet:
+                    return "Outfit Glamours";
+                case SettingSubCategory.CofferLoot:
+                    return "Coffer Loot";
             }
             return settingSubCategory.ToString();
         }