+++ InventoryTools/Logic/Settings/Abstract/SettingSubCategory.cs
@@ -30,5 +30,7 @@ namespace InventoryTools.Logic.Settings.Abstract
         UseInformation,
         AcquisitionTracker,
         IngredientPatch,
+        GlamourReadySet,
+        CofferLoot,
     }
 }